refactor: fold the deployments module into core/site
The merge kept the full-stack deploy code in `core/deployments/`, but main
had already settled this: deployments are a transport of the site module, not
a module of their own, and they live in `core/site/`. Follow main.

`core/deployments/` is gone. One flow per file — `full-stack.ts` (Workers, the
cf arm), `static-site.ts` (deployments-API static, the s3 arm), `deploy.ts`
(legacy tar.gz) — over shared `manifest.ts`, `upload.ts`, `modules.ts`,
`wrangler-config.ts`, and `git-hash.ts`. The deployment requests and schemas
merge into the module's existing `api.ts` / `schema.ts` next to the tar.gz
upload, as main has them. `deploy-app.ts` stays the transport picker.

The Workers flow is `full-stack.ts` rather than `deploy.ts` because that name
is already the legacy tar.gz path; it reads as a pair with `static-site.ts`.
Unit tests follow the same naming: `tests/core/site-*.spec.ts`.

No behavior change — moves, import rewrites, and the barrel/doc updates that
follow from them. While merging the two `api.ts` files, the worker-module read
switched to the existing `@/core/utils/fs.js` helper instead of a second raw
`readFile` binding, for typed FileNotFound/FileRead errors.

typecheck, lint, and knip clean; the 8 deploy spec files pass (77 tests).
