--- packages/cli/src/core/deployments/index.ts
@@ -1,9 +1,0 @@
-export * from "./api.js";
-export * from "./deploy.js";
-export * from "./git-hash.js";
-export * from "./manifest.js";
-export * from "./modules.js";
-export * from "./schema.js";
-export * from "./static-site.js";
-export * from "./upload.js";
-export * from "./wrangler-config.js";
