+++ packages/cli/src/core/index.ts
@@ -2,7 +2,6 @@ export * from "./auth/index.js";
 export * from "./clients/index.js";
 export * from "./config.js";
 export * from "./consts.js";
-export * from "./deployments/index.js";
 export * from "./errors.js";
 export * from "./project/index.js";
 export * from "./resources/index.js";
