+++ packages/cli/src/core/site/full-stack.ts
