+++ packages/cli/src/core/site/git-hash.ts
