+++ packages/cli/src/core/site/index.ts
@@ -2,4 +2,11 @@ export * from "./api.js";
 export * from "./config.js";
 export * from "./deploy.js";
 export * from "./deploy-app.js";
+export * from "./full-stack.js";
+export * from "./git-hash.js";
+export * from "./manifest.js";
+export * from "./modules.js";
 export * from "./schema.js";
+export * from "./static-site.js";
+export * from "./upload.js";
+export * from "./wrangler-config.js";
