+++ packages/cli/src/core/site/manifest.ts
