+++ packages/cli/src/core/site/static-site.ts
